Add `types` folder and moved all models/relevant types
