+++ client/types/index.ts
@@ -1,0 +1,6 @@
+export * from "./common";
+export * from "./company";
+export * from "./gig";
+export * from "./job";
+export * from "./link";
+export * from "./user";
