+++ client/types/job.ts
@@ -1,44 +1,45 @@
 import {
   Application,
-  ApplicationStatus,
   CreatedBy,
   PostingStatus,
   PrimaryKey,
   WithTimestamp,
-} from "./common.model";
+} from "./common";
 
-export type Gig = PrimaryKey & {
+export type Job = PrimaryKey & {
+  company_id: string;
   title: string;
   description: string;
   position?: string;
+  opportunity?: string;
   featured_photo_url?: string;
-  experience_level?: string;
   status?: PostingStatus;
 } & WithTimestamp &
   CreatedBy;
 
-export type GigForm = Pick<
-  Gig,
-  | "title"
+export type JobForm = Pick<
+  Job,
+  | "company_id"
   | "description"
+  | "title"
   | "position"
+  | "opportunity"
   | "featured_photo_url"
-  | "experience_level"
   | "status"
 >;
 
-export type GigCategory = {
-  gig_id: string;
+export type JobCategory = {
+  job_id: string;
   category_id: string;
 } & WithTimestamp;
 
-export type GigApplication = PrimaryKey & {
-  gig_id: string;
+export type JobApplication = PrimaryKey & {
+  job_id: string;
 } & Application &
   WithTimestamp &
   CreatedBy;
 
-export type GigApplicationForm = Pick<
-  GigApplication,
-  "gig_id" | "pitch" | "content" | "status"
+export type JobApplicationForm = Pick<
+  JobApplication,
+  "job_id" | "pitch" | "content" | "status"
 >;
